+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## [1.2.0] - 07 Oct 2019 
+
+- Add ProgressDialog
+- Add ProgressDialog with Future
+- Set your dialog look with DialogStyle
+
 ## [1.1.0] - 03 Oct 2019 
 
 - Fix overflow actions